+++ assets/javascripts/kitten/components/cards/crowdfunding-card/test.js
@@ -2,7 +2,6 @@ import React from 'react'
 import renderer from 'react-test-renderer'
 import 'jest-styled-components'
 import { CrowdfundingCard } from './index'
-import { CrowdfundingCardWidget } from './widget'
 
 const createMockMediaMatcher = matches => () => ({
   matches,
@@ -44,35 +43,6 @@ describe('<CrowdfundingCard />', () => {
             info3="Lorem ipsum…"
             progress={42}
             stretch={true}
-          />,
-        )
-        .toJSON()
-    })
-
-    it('matches with snapshot', () => {
-      expect(component).toMatchSnapshot()
-    })
-  })
-
-  describe('CrowdfundingCardWidget props', () => {
-    beforeEach(() => {
-      window.matchMedia = createMockMediaMatcher(false)
-
-      component = renderer
-        .create(
-          <CrowdfundingCardWidget
-            ownerTitle="Owner title"
-            ownerDescription="Owner description"
-            imageProps={{ src: '#image', alt: '' }}
-            avatarProps={{ src: '#image', alt: '' }}
-            state="Sur"
-            loading={false}
-            title="Lorem ipsum…"
-            subtitle="Sed ut perspiciatis unde omnis iste natus error Lorem"
-            titleTruncate
-            subTitleTruncate
-            buttonText="Lorem ipsum"
-            dayCounter={null}
           />,
         )
         .toJSON()
@@ -127,7 +97,7 @@ describe('<CrowdfundingCard />', () => {
 
       component = renderer
         .create(
-          <CrowdfundingCardWidget
+          <CrowdfundingCard
             ownerTitle="Owner title"
             ownerDescription="Owner description"
             imageProps={{ src: '#image', alt: '' }}
@@ -157,7 +127,7 @@ describe('<CrowdfundingCard />', () => {
 
       component = renderer
         .create(
-          <CrowdfundingCardWidget
+          <CrowdfundingCard
             ownerTitle="Owner title"
             ownerDescription="Owner description"
             imageProps={{ src: '#image', alt: '' }}